CQRS with MinimalAPI for the project Table

+++ TeamcollborationHub.server/Repositories/UserRepository/UserRepository.cs
@@ -9,7 +9,7 @@ namespace TeamcollborationHub.server.Repositories.UserRepository;
 /// </summary>
 public class UserRepository : IUserRepository
 {
-    
+
     /// <summary>
     /// an instance of TDBContext to interact with the database
     /// </summary>
@@ -35,7 +35,7 @@ public class UserRepository : IUserRepository
     /// <returns>return the created user , good for testing </returns>
     public async Task<User?> CreateUser(User user)
     {
-        
+
         _context.Users.Add(user);
         await _context.SaveChangesAsync();
         return user;
@@ -75,18 +75,19 @@ public class UserRepository : IUserRepository
     /// </summary>
     /// <param name="id"></param>
     /// <returns></returns>
-    public async Task<User> deleteUser(int id)
+    public async Task<User?> DeleteUser(int id)
     {
         var user = await _context.FindAsync<User>(id);
-        if (user == null) return user!;
+        if (user is null) return null;
         _context.Users.Remove(user);
         await _context.SaveChangesAsync();
         return user;
     }
 
-    public async Task<User> deleteUser(string email)
+    public async Task<User?> DeleteUser(string email)
     {
-        var result =await _context.Users.SingleOrDefaultAsync(u => u.Email == email);
+        var result = await _context.Users.SingleOrDefaultAsync(u => u.Email == email) ;
+        if (result is null) return null;
         _context.Users.Remove(result);
         await _context.SaveChangesAsync();
         return result;
